add utility functions for manually controlling and setting odd arm solutions

+++ src/modules/utils/panel/screens/WatchScreen.cpp
@@ -65,7 +65,7 @@ void WatchScreen::on_enter()
     get_current_status();
     get_current_pos(this->pos);
     get_sd_play_info();
-    this->current_speed = lround(get_current_speed());
+    this->current_speed = lroundf(get_current_speed());
     this->refresh_screen(false);
     THEPANEL->enter_control_mode(1, 0.5);
     THEPANEL->set_control_value(this->current_speed);
@@ -122,7 +122,7 @@ void WatchScreen::on_refresh()
             this->speed_changed = false;
         } else if (!this->issue_change_speed) { // change still queued
             // read it in case it was changed via M220
-            this->current_speed = lround(get_current_speed());
+            this->current_speed = lroundf(get_current_speed());
             THEPANEL->set_control_value(this->current_speed);
             THEPANEL->reset_counter();
         }
